Always set Zoom Fit mode at application startup.
spicview tryes to determine the best zoom mode (Fit or 100%) depending on available screen size and image size.
It does non work well, since the mode gets unpredictable and it looks like an application bug from user's point of view.
So this feature is removed now.

The mode selection code is removed. Fit mode is unconditionally enabled at application startup.
The files working-area.[hc], that contains the implementation of get_working_area(), are removed as well.
X11 CLFAGS and LIBS are not checked in ./configure any more, sinse working-area.c was the only user of them.