"Save As" should not reload the image list from the new file location.
"Save As" works as supposed: it saves the current file in a new location and switches to that location, reloading the file list.
This behaviour, though correct, can be confusing, especially given the fact that we do not display the full file path anywhere in the UI.
So, it is better to get rid of this feature.

The name of the action in the UI is changed to "Save a Copy" as well.