Pack the application resources into compiled-in GResource container.
The application icon, the fallback icons and the UI files are now linked directly into the application executable.
The only external resource required by the applcation to run is the gettext catalog.

Also, install the application icon into the hicolor directory. It was installed into $(datadir)/pixmaps for some reason.